Update to handle now redundant block
