+++ pymagicc/definitions/magicc_dattype_regionmode_regions.csv
@@ -14,10 +14,20 @@ SCEN7 files only. Order doesn't matter and any other regions in the file are ign
 SCEN7 files only. Order doesn't matter and any other regions in the file are ignored. If you give this config to a non-SCEN7 file then MAGICC will assume it's global data based off the fact that it's only a single column and THISFILE_DATTYPE ~= 'REGIONDATA',SCEN7,WORLD,GL,,,,,,,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 5 columns and do nothing.,SCEN7,SRES,OECD90,REF,ASIA,ALM,,,,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCP,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCP,R6OECD,R6REF,R6ASIA,R6MAF,R6LAM,,,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCP,WORLD,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCP,WORLD,R6OECD,R6REF,R6ASIA,R6MAF,R6LAM,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,BUNKERS,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,R5OECD90,R5REF,R5ASIA,R5MAF,R5LAM,BUNKERS,,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,BUNKERS,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,WORLD,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,BUNKERS,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,WORLD,R5OECD90,R5REF,R5ASIA,R5MAF,R5LAM,BUNKERS,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,WORLD,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,BUNKERS,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,AVIATION,SHIPPING,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,R5OECD90,R5REF,R5ASIA,R5MAF,R5LAM,AVIATION,SHIPPING,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,AVIATION,SHIPPING,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,WORLD,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,AVIATION,SHIPPING,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,WORLD,R5OECD90,R5REF,R5ASIA,R5MAF,R5LAM,AVIATION,SHIPPING,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,WORLD,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,AVIATION,SHIPPING,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,GLOBAL,,,,,,,,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,WORLD,,,,,,,,,,,,,,,,,,,
@@ -25,3 +35,5 @@ Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,NHLAND,SHLAND,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,GLOBAL,NHLAND,SHLAND,,,,,,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,NHOCEAN,NHLAND,SHOCEAN,SHLAND,,,,,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,HEMISPHERIC,NH,SH,,,,,,,,,,,,,,,,,,
+
+
